rename: 기존의 연관상품 조회와 관련한 파일명 related 추가
